requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
html2text>=2020.1.16
python-dotenv>=1.0.0
